changed implementation into datamodel and metadata
